+++ test/api.types.ts
@@ -1,8 +1,6 @@
-import * as parse from '../lib/index';
-import {CastingContext, Options, Parser} from '../lib/index';
-// import {} from 'should';
-// if you used the '@types/mocha' method to install mocha type definitions, uncomment the following line
-// import 'mocha';
+
+import * as parse from '../lib/index'
+import {CastingContext, Info, Options, Parser} from '../lib/index'
 
 describe('API Types', () => {
   
@@ -21,151 +19,184 @@ describe('API Types', () => {
         'skip_lines_with_error', 'to', 'to_line', 'trim'
       ])
     })
+      
+    it('Expose info', () => {
+      const parser: Parser = parse()
+      const info: Info = parser.info
+      const keys: any = Object.keys(info)
+      keys.sort().should.eql([
+        'empty_line_count', 'lines', 'records', 'skipped_line_count'
+      ])
+    })
+    
+  })
+  
+  describe('Info', () => {
+    
+    it('empty_line_count', () => {
+      const info: Info = {}
+      info.empty_line_count = 1
+    })
+      
+    it('lines', () => {
+      const info: Info = {}
+      info.lines = 1
+    })
+    
+    it('records', () => {
+      const info: Info = {}
+      info.records = 1
+    })
+      
+    it('skipped_line_count', () => {
+      const info: Info = {}
+      info.skipped_line_count = 1
+    })
     
   })
   
   describe('Options', () => {
     
     it('cast', () => {
-      const options: Options = {};
-      options.cast = true;
-      options.cast = () => {};
-    });
+      const options: Options = {}
+      options.cast = true
+      options.cast = () => {}
+    })
       
     it('cast_date', () => {
-      const options: Options = {};
-      options.cast_date = true;
-    });
+      const options: Options = {}
+      options.cast_date = true
+    })
       
     it('columns', () => {
-      const options: Options = {};
-      options.columns = true;
-      options.columns = [];
+      const options: Options = {}
+      options.columns = true
+      options.columns = []
       options.columns = (record: string[]) => {
         const val: boolean = record.length > 1
         return val
-      };
+      }
       options.columns = (record: string[]) => {
         const fields: string[] = record.map( (field: string) => {
           return field.toUpperCase()
         })
         return fields
-      };
-    });
+      }
+    })
       
     it('comment', () => {
-      const options: Options = {};
-      options.comment = '\\';
-    });
+      const options: Options = {}
+      options.comment = '\\'
+    })
     
     it('delimiter', () => {
-      const options: Options = {};
-      options.delimiter = ':';
-      options.delimiter = Buffer.from(':');
-    });
+      const options: Options = {}
+      options.delimiter = ':'
+      options.delimiter = Buffer.from(':')
+    })
     
     it('escape', () => {
-      const options: Options = {};
-      options.escape = ':';
-      options.escape = Buffer.from('"');
-    });
+      const options: Options = {}
+      options.escape = ':'
+      options.escape = Buffer.from('"')
+    })
     
     it('from', () => {
-      const options: Options = {};
-      options.from = 10;
-    });
+      const options: Options = {}
+      options.from = 10
+    })
     
     it('from_line', () => {
-      const options: Options = {};
-      options.from_line = 10;
-    });
+      const options: Options = {}
+      options.from_line = 10
+    })
     
     it('ltrim', () => {
-      const options: Options = {};
-      options.ltrim = true;
-    });
+      const options: Options = {}
+      options.ltrim = true
+    })
     
     it('max_record_size', () => {
-      const options: Options = {};
-      options.max_record_size = 100;
-    });
+      const options: Options = {}
+      options.max_record_size = 100
+    })
     
     it('objname', () => {
-      const options: Options = {};
-      options.objname = 'name';
-    });
+      const options: Options = {}
+      options.objname = 'name'
+    })
     
     it('quote', () => {
-      const options: Options = {};
-      options.quote = '"';
-      options.quote = true;
-      options.quote = Buffer.from('"');
-    });
+      const options: Options = {}
+      options.quote = '"'
+      options.quote = true
+      options.quote = Buffer.from('"')
+    })
     
     it('raw', () => {
-      const options: Options = {};
-      options.raw = true;
-    });
+      const options: Options = {}
+      options.raw = true
+    })
     
     it('relax', () => {
-      const options: Options = {};
-      options.relax = true;
-    });
+      const options: Options = {}
+      options.relax = true
+    })
     
     it('relax_column_count', () => {
-      const options: Options = {};
-      options.relax_column_count = true;
-    });
+      const options: Options = {}
+      options.relax_column_count = true
+    })
     
     it('record_delimiter', () => {
-      const options: Options = {};
-      options.record_delimiter = '\n';
-      options.record_delimiter = ['\n'];
-      options.record_delimiter = Buffer.from('\n');
-      options.record_delimiter = [Buffer.from('\n')];
-    });
+      const options: Options = {}
+      options.record_delimiter = '\n'
+      options.record_delimiter = ['\n']
+      options.record_delimiter = Buffer.from('\n')
+      options.record_delimiter = [Buffer.from('\n')]
+    })
     
     it('rtrim', () => {
-      const options: Options = {};
-      options.rtrim = true;
-    });
+      const options: Options = {}
+      options.rtrim = true
+    })
     
     it('skip_empty_lines', () => {
-      const options: Options = {};
-      options.skip_empty_lines = true;
-    });
+      const options: Options = {}
+      options.skip_empty_lines = true
+    })
     
     it('skip_empty_lines', () => {
-      const options: Options = {};
-      options.skip_empty_lines = true;
-    });
+      const options: Options = {}
+      options.skip_empty_lines = true
+    })
     
     it('skip_lines_with_empty_values', () => {
-      const options: Options = {};
-      options.skip_lines_with_empty_values = true;
-    });
+      const options: Options = {}
+      options.skip_lines_with_empty_values = true
+    })
     
     it('skip_lines_with_error', () => {
-      const options: Options = {};
-      options.skip_empty_lines = true;
-    });
+      const options: Options = {}
+      options.skip_empty_lines = true
+    })
     
     it('to', () => {
-      const options: Options = {};
-      options.to = 10;
-    });
+      const options: Options = {}
+      options.to = 10
+    })
     
     it('to_line', () => {
-      const options: Options = {};
-      options.to_line = 10;
-    });
+      const options: Options = {}
+      options.to_line = 10
+    })
     
     it('trim', () => {
-      const options: Options = {};
-      options.trim = true;
-    });
+      const options: Options = {}
+      options.trim = true
+    })
     
-  });
+  })
   
   describe('CastingContext', () => {
     
@@ -187,4 +218,4 @@ describe('API Types', () => {
     })
   })
   
-});
+})